Make page loader recognize URL with #
This allows cross-page #links to work.

Because of dynamic loading, the location of a #link will likely change
from where it is when the page initially loads. To compensate, I’ve put
an interval that will re-adjust the scroll position a few times. A
fully correct solution would require each page to signal when it was
stable, which would require significant work on each page.