Load images using old API
API was unable to be updated to include image urls for this release,
so we've had to fallback to loading images with additional roundtrip
requests.

CMSPageController will automatically traverse the JSON and find any
`image` property with a number and create a separate call to fetch the
image's url and insert it in place of the number.