Form fixes (#503)
* Form fixes

* Use detach instead of onClose

The select handler tries to tie itself to onClose, which is
getting broken when the view also defines onClose. This might
be something to find a more elegant solution to later.

* Make Give and Adoption validate like Bulk Order

* Form fixes

Make all forms validate like Bulk Order

Also:
Use detach instead of onClose

The select handler tries to tie itself to onClose, which is
getting broken when the view also defines onClose. This might
be something to find a more elegant solution to later.