Fix request for nonexistent math.svg
The generic css for the banners sets background image for .book and
.second-book to be ${subject}.svg. There is no math.svg. Although the
css overrides that generic css, we are still getting requests. I have
put in an if-directive to prevent the setting of the background image
in the generic css when subject is math.

layout for header and details
Revert "layout for header and details"
Restore old get-this-title and rename new one
Update Fontawesome icons
Fix layout issues for adoption form buttons and confirmation header
Comp copy dialogs display