Fix iframe as target of forms issue (#1575)
* Fix duplicate testimonial form on close and reopen

Was displaying a new form every time the window opened, but only removing the previous form on submit. So closing the window caused duplicate form.

* Refactor the Contact form

Create a salesforce-form component for forms that target an iframe and go to a confirmation
page after submit.

If assembled in `onDataLoaded`, the form will not target the iframe. if assembled in `onLoaded`, it will.

* Update version