Port models to ts (#2701)
* myopenstax-user

* wip - reviews

* renewals

* reviews -> partner-info

The ratings functionality is no longer used; took it out to simplify the codebase.
Renamed to be clearer about purpose.

* Quiet console.info in page-data-utils tests

* rex-release

cnxFetch did not need to exist anymore

* savings, salesforce-partners

* revise window-settings

Simplifies the call, but for some reason link.ts was not recognizing that the exported settings function was a function, so it had to be worked around.

* sfapi

* table-of-contents-html

* use-school-suggestion-list

* Extract accounts-model from usermodel

usermodel is hairy

* usermodel

* Prettier

* Cover queryById case

* Tweak return type of accounts-model load

Unnecessary ?